maana-nltk/requirements.txt
aio-pika==2.3.0
aiohttp==3.0.9
aiohttp-cors==0.7.0
aiohttp-graphql==1.0.0
async-timeout==2.0.1
asyncio==3.4.3
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
future==0.16.0
gevent==1.2.2
gevent-websocket==0.10.1
graphene==2.0.1
graphql-core==2.0
graphql-relay==0.4.5
graphql-server-core==1.0.dev20170322001
graphql-ws==0.2.0
greenlet==0.4.13
idna==2.6
idna-ssl==1.0.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
multidict==4.1.0
pika==0.11.2
promise==2.1
pytest-runner==4.0
python-dotenv==0.8.2
redis==2.10.6
requests==2.18.4
Rx==1.6.1
shortuuid==0.5.0
six==1.11.0
typing==3.6.4
ujson==1.35
urllib3==1.22
uvloop==0.9.1
Werkzeug==0.14.1
yarl==1.1.1
nltk==3.2.5